AIR-4497 Disabled code which causes pyarmnn cross-compile failures on yocto/arm64
There are 2 issues which cause cross-compilation to fail in the Yocto environment, and the upstream code
is not yet supporting cross-compiliation, thus some code needs disabled to work around this.

a) Checking the Arm NN version requires to call GetVersion from a lib targetted for the same platform
as it is being compiled on, which is not available when cross-compiling.
b) The Python setuptools.Extension module appends the argument -R to GCC to define explicit link paths,
which is unknown to ARM GCC, so we only declare the libs to link and expect, that they can be found in
default link dirs.

Signed-off-by: Pavel Macenauer <pavel.macenauer@nxp.com>